+++ fugue/src/main/java/io/atlassian/fugue/Try.java
@@ -10,9 +10,9 @@ import java.util.function.BinaryOperator;
 import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.stream.Collector;
+import java.util.stream.Collectors;
 
 import static io.atlassian.fugue.Suppliers.memoize;
-import static java.util.Collections.unmodifiableList;
 import static java.util.Objects.requireNonNull;
 import static java.util.function.Function.identity;
 
@@ -67,19 +67,36 @@ public abstract class Try<A> {
    * @return a success wrapping all of the values if all of the arguments were a
    * success, otherwise this returns the first failure
    */
-  public static <A> Try<Iterable<A>> sequence(Iterable<Try<A>> trys) {
-    final ArrayList<A> ts = new ArrayList<>();
-    for (final Try<A> t : trys) {
+  public static <A> Try<Iterable<A>> sequence(final Iterable<Try<A>> trys) {
+    return sequence(trys, Collectors.collectingAndThen(Collectors.toList(), Collections::unmodifiableList));
+  }
+
+  /**
+   * Returns a success wrapping all of the values if all of the arguments were a
+   * success, otherwise this returns the first failure
+   *
+   * @param trys an iterable of try values
+   * @param collector result collector
+   * @param <T> The success type
+   * @param <A> The intermediate accumulator type
+   * @param <R> The result type
+   * @return a success wrapping all of the values if all of the arguments were a
+   * success, otherwise this returns the first failure
+   * @since 4.6.0
+   */
+  public static <T, A, R> Try<R> sequence(final Iterable<Try<T>> trys, final Collector<T, A, R> collector) {
+    A accumulator = collector.supplier().get();
+    for (final Try<T> t : trys) {
       if (t.isFailure()) {
         return new Failure<>(t.fold(identity(), x -> {
           throw new NoSuchElementException();
         }));
       }
-      ts.add(t.fold(f -> {
+      collector.accumulator().accept(accumulator, t.fold(f -> {
         throw new NoSuchElementException();
       }, identity()));
     }
-    return new Success<>(unmodifiableList(ts));
+    return new Success<>(collector.finisher().apply(accumulator));
   }
 
   public static <A> SequenceCollector<A> sequenceCollector() {
@@ -216,12 +233,13 @@ public abstract class Try<A> {
   public abstract Try<A> recover(Function<? super Exception, A> f);
 
   /**
-   * Applies the given function `f` if this is a `Failure` with certain exception type otherwise leaves this
-   * unchanged. This is like map for exceptions types.
+   * Applies the given function `f` if this is a `Failure` with certain
+   * exception type otherwise leaves this unchanged. This is like map for
+   * exceptions types.
    *
    * @param exceptionType exception class
-   * @param f             the function to apply
-   * @param <X>           exception type
+   * @param f the function to apply
+   * @param <X> exception type
    * @return `f` applied to the `Failure`, otherwise returns this if this is a
    * `Success` or the exception does not match the exception type.
    */
@@ -238,12 +256,12 @@ public abstract class Try<A> {
   public abstract Try<A> recoverWith(Function<? super Exception, Try<A>> f);
 
   /**
-   * Binds the given function across certain exception type if it is one, otherwise
-   * this unchanged. This is like flatmap for exceptions types.
+   * Binds the given function across certain exception type if it is one,
+   * otherwise this unchanged. This is like flatmap for exceptions types.
    *
    * @param exceptionType exception class
-   * @param f             the function to apply
-   * @param <X>           exception type
+   * @param f the function to apply
+   * @param <X> exception type
    * @return A new Try value after binding with the function applied if this is
    * a `Failure`, otherwise returns this if this is a `Success` or the exception
    * does not match the exception type.
@@ -316,8 +334,7 @@ public abstract class Try<A> {
       return Checked.of(() -> f.apply(e));
     }
 
-    @SuppressWarnings("unchecked")
-    @Override public <X extends Exception> Try<A> recover(final Class<X> exceptionType, final Function<? super X, A> f) {
+    @SuppressWarnings("unchecked") @Override public <X extends Exception> Try<A> recover(final Class<X> exceptionType, final Function<? super X, A> f) {
       return exceptionType.isAssignableFrom(e.getClass()) ? Checked.of(() -> f.apply((X) e)) : this;
     }
 
@@ -325,8 +342,7 @@ public abstract class Try<A> {
       return f.apply(e);
     }
 
-    @SuppressWarnings("unchecked")
-    @Override public <X extends Exception> Try<A> recoverWith(Class<X> exceptionType, Function<? super X, Try<A>> f) {
+    @SuppressWarnings("unchecked") @Override public <X extends Exception> Try<A> recoverWith(Class<X> exceptionType, Function<? super X, Try<A>> f) {
       return exceptionType.isAssignableFrom(e.getClass()) ? f.apply((X) e) : this;
     }