+++ fugue/src/test/java/io/atlassian/fugue/TryTest.java
@@ -5,6 +5,7 @@ import org.junit.Test;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.NoSuchElementException;
+import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
 import java.util.stream.IntStream;
@@ -12,7 +13,9 @@ import java.util.stream.Stream;
 
 import static io.atlassian.fugue.Functions.identity;
 import static java.util.stream.Collectors.toList;
+import static java.util.stream.Collectors.toSet;
 import static org.hamcrest.CoreMatchers.instanceOf;
+import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
 
@@ -164,6 +167,18 @@ public class TryTest {
     assertThat(e.getMessage(), is("FIRST 1"));
 
     assertThat(called.get(), is(1));
+  }
+
+  @Test public void sequenceReturnsValuesFromAllSuccessesWithCustomCollector() {
+    Try<Set<Integer>> result = Try.sequence(IntStream.range(0, 10).map(i -> i % 4)
+            .mapToObj(i -> Checked.of(() -> i))
+            .collect(toList()), toSet());
+
+    assertThat(result.isSuccess(), is(true));
+    Set<Integer> vals = result.fold(f -> {
+      throw new NoSuchElementException();
+    }, identity());
+    assertThat(vals, contains(0, 1, 2, 3));
   }
 
   @Test public void flattenNestedSuccess() {